Fix the counter growth discontinuous in raft mode
