+++ hugegraph-core/src/main/java/com/baidu/hugegraph/backend/store/raft/StoreClosure.java
@@ -19,80 +19,23 @@
 
 package com.baidu.hugegraph.backend.store.raft;
 
-import static com.baidu.hugegraph.backend.store.raft.RaftSharedContext.WAIT_RAFT_LOG_TIMEOUT;
-
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
-import java.util.function.Supplier;
-
 import org.slf4j.Logger;
 
-import com.alipay.sofa.jraft.Closure;
-import com.alipay.sofa.jraft.Status;
-import com.baidu.hugegraph.backend.BackendException;
 import com.baidu.hugegraph.util.E;
 import com.baidu.hugegraph.util.Log;
 
-public class StoreClosure implements Closure {
+public class StoreClosure extends RaftClosure {
 
     private static final Logger LOG = Log.logger(StoreClosure.class);
 
     private final StoreCommand command;
-    private final CompletableFuture<RaftResult> future;
 
     public StoreClosure(StoreCommand command) {
         E.checkNotNull(command, "store command");
         this.command = command;
-        this.future = new CompletableFuture<>();
     }
 
     public StoreCommand command() {
         return this.command;
-    }
-
-    public Object waitFinished() throws Throwable {
-        RaftResult result = this.get();
-        if (result.status().isOk()) {
-            return result.callback().get();
-        } else {
-            throw result.exception();
-        }
-    }
-
-    public Status status() {
-        return this.get().status();
-    }
-
-    private RaftResult get() {
-        try {
-            return this.future.get(WAIT_RAFT_LOG_TIMEOUT, TimeUnit.MILLISECONDS);
-        } catch (InterruptedException | ExecutionException e) {
-            throw new BackendException("ExecutionException", e);
-        } catch (TimeoutException e) {
-            throw new BackendException("Wait store command '%s' timeout",
-                                       this.command.action());
-        }
-    }
-
-    public void complete(Status status, Supplier<Object> callback) {
-        this.future.complete(new RaftResult(status, callback));
-    }
-
-    public void failure(Status status, Throwable exception) {
-        this.future.complete(new RaftResult(status, exception));
-    }
-
-    @Override
-    public void run(Status status) {
-        if (status.isOk()) {
-            this.complete(status, () -> null);
-        } else {
-            LOG.error("Failed to apply command: {}", status);
-            String msg = "Failed to apply command in raft node with error : " +
-                         status.getErrorMsg();
-            this.failure(status, new BackendException(msg));
-        }
     }
 }
